user-service/requirements.txt
Flask==2.3.2  
Flask-Dance==6.2.0  
Flask-SQLAlchemy==3.1.1  
Flask-Migrate==4.0.5  
Flask-Cors==4.0.0  
Flask-JWT-Extended==4.6.0  
python-dotenv==1.0.1  
PyJWT==2.8.0  
psycopg2-binary==2.9.9  
requests==2.31.0  
oauthlib==3.2.2
kafka-python-ng==2.1.0 # Added for Kafka integration